[refactor]postのフォルダ整理
